+++ DSCResources/MSFT_xSQLServerSetup/MSFT_xSQLServerSetup.psm1
@@ -387,7 +387,7 @@ function Get-TargetResource
     .PARAMETER InstanceID
         SQL instance ID, if different from InstanceName.
 
-    .PARAMETER PID
+    .PARAMETER ProductKey
         Product key for licensed installations.
 
     .PARAMETER UpdateEnabled
@@ -530,7 +530,7 @@ function Set-TargetResource
         $InstanceID,
 
         [System.String]
-        $PID,
+        $ProductKey,
 
         [System.String]
         $UpdateEnabled,
@@ -888,7 +888,7 @@ function Set-TargetResource
         'UpdateEnabled',
         'UpdateSource',
         'Features',
-        'PID',
+        'ProductKey',
         'SQMReporting',
         'ErrorReporting',
         'InstallSharedDir',
@@ -989,7 +989,14 @@ function Set-TargetResource
     # Automatically include any additional arguments
     foreach ($argument in $argumentVars)
     {
-        $setupArguments += @{ $argument = (Get-Variable -Name $argument -ValueOnly) }
+        if($argument -eq 'ProductKey') 
+        {
+            $setupArguments += @{ 'PID' = (Get-Variable -Name $argument -ValueOnly) }
+        }
+        else
+        {
+            $setupArguments += @{ $argument = (Get-Variable -Name $argument -ValueOnly) }
+        }
     }
 
     # Build the argument string to be passed to setup
@@ -1034,9 +1041,9 @@ function Set-TargetResource
         $log = $log.Replace($SAPwd.GetNetworkCredential().Password,"********")
     }
 
-    if ($PID -ne "")
+    if ($ProductKey -ne "")
     {
-        $log = $log.Replace($PID,"*****-*****-*****-*****-*****")
+        $log = $log.Replace($ProductKey,"*****-*****-*****-*****-*****")
     }
 
     $logVars = @('AgtSvcAccount', 'SQLSvcAccount', 'FTSvcAccount', 'RSSvcAccount', 'ASSvcAccount','ISSvcAccount')
@@ -1112,7 +1119,7 @@ function Set-TargetResource
     .PARAMETER InstanceID
         SQL instance ID, if different from InstanceName.
 
-    .PARAMETER PID
+    .PARAMETER ProductKey
         Product key for licensed installations.
 
     .PARAMETER UpdateEnabled
@@ -1254,7 +1261,7 @@ function Test-TargetResource
         $InstanceID,
 
         [System.String]
-        $PID,
+        $ProductKey,
 
         [System.String]
         $UpdateEnabled,