+++ CHANGELOG.md
@@ -35,6 +35,7 @@
       - Function `ResolvePath` has been replaced with `[Environment]::ExpandEnvironmentVariables($SourcePath)` so that environment variables still can be used in Source Path.
       - Function `NetUse` has been replaced with `New-SmbMapping` and `Remove-SmbMapping`.
   - Renamed function `GetSQLVersion` to `Get-SqlMajorVersion`.
+  - BREAKING CHANGE: Renamed parameter PID to ProductKey to avoid collision with automatic variable $PID
 - Changes to xSQLServerScript
   - All credential parameters now also has the type [System.Management.Automation.Credential()] to better work with PowerShell 4.0.
   - It is now possible to configure two instances on the same node, with the same script.